requirements.txt
flask~=3.1.1
pytest~=8.3.5
pytest-cov
flask_sqlalchemy
Werkzeug~=3.1.3
locust~=2.37.4
requests~=2.32.3
selenium~=4.33.0
python-dotenv~=1.1.0
Flask-Migrate~=4.1.0
MarkupSafe~=3.0.2
PyYAML~=6.0.2
beautifulsoup4~=4.13.4